+++ examples/service.customer.js
@@ -51,7 +51,7 @@
 			}
 		}), im = ui.im;
 
-		im.setUser({"id":"admin","uid":"admin","nick":"\u7ba1\u7406\u5458","group_id":"1","pic_url":"http:\/\/www.gravatar.com\/avatar\/?s=50","default_pic_url":"http:\/\/www.gravatar.com\/avatar\/?s=50","show":"unavailable","status":null,"url":"http:\/\/www.gravatar.com\/"});
+		im.setUser({"id":"admin","uid":"admin","nick":"\u7ba1\u7406\u5458","group_id":"1","avatar":"http:\/\/www.gravatar.com\/avatar\/?s=50","default_avatar":"http:\/\/www.gravatar.com\/avatar\/?s=50","show":"unavailable","status":null,"url":"http:\/\/www.gravatar.com\/"});
 
 		ui.addApp("buddy", {
 			is_login: true,